begin filters page

+++ app/src/main/java/live/noxbox/Configuration.java
@@ -3,7 +3,7 @@ package live.noxbox;
 public interface Configuration {
 
     int ACCEPTANCE_TIMEOUT_IN_SECONDS = 30;
-    int RADIUS_IN_METERS = 50 * 1000;
+    double RADIUS_IN_METERS = 50 * 1000;
     int MIN_RATE_IN_PERCENTAGE = 90;
     String CURRENCY = "Waves";
     int DAY_TIME_START_IN_HOURS = 10;
